Move get_args_from_envvars() from environment to compilers
